audio перенесен выше ( в компонент Bar)
